backend/requirements.txt
python-dotenv
requests-cache
pytest
pytest-mock
yfinance
pydantic
google-genai